First implementation of a cache for static files
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@2127 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserver/source/myserver.cpp
@@ -196,8 +196,8 @@ int main (int argn, char **argv)
   };
 	
 	{
-		int pathLen;
-		u_long len;
+		u_int pathLen;
+		u_int len;
 		bool differentCwd = false;
 
 		pathLen = strlen(argv[0]);
